don't use pidfile for tomcat on el7 (SOFTWARE-1996)
on el7, there is always a /var/run/tomcat.pid present, but it doesn't
actually seem to get used by systemd for maintaining the service, and
thus osg-test always thinks it's running and doesn't try to start
tomcat.  there doesn't seem to be a real pidfile for tomcat on el7.

git-svn-id: https://vdt.cs.wisc.edu/svn/software/osg-test/trunk@21883 4e558342-562e-0410-864c-e07659590f8c