In el7, kill the pbs_server process used to create the db (SOFTWARE-2130)
Hey, it looks ugly, but check out the create() function in the el6
init.d script, and you'll see the even more confusing version that it
came from.

Anyways, it seems that the call to 'qterm' is necessary to kill the
pbs_server used to create the pbs db before service start (or stop) will
work.  The el6 init.d script's create() function takes care of this but,
as mentioned elsewhere, there isn't an equivalent for el7/systemd...

git-svn-id: https://vdt.cs.wisc.edu/svn/software/osg-test/trunk@22192 4e558342-562e-0410-864c-e07659590f8c